+++ packages/edit-site/src/components/post-list/index.js
@@ -296,6 +296,8 @@ export default function PostList( { postType } ) {
 			_embed: 'author',
 			order: view.sort?.direction,
 			orderby: view.sort?.field,
+			orderby_hierarchy:
+				view.type === 'table' && !! view.layout?.hierarchicalSort,
 			search: view.search,
 			...filters,
 		};